requirements.txt
ping3==4.0.4
paho-mqtt==1.6.1
